add support for cypher directive with scalar result
resolves #85
