+++ core/src/test/kotlin/org/neo4j/graphql/IssuesTests.kt
@@ -1,0 +1,10 @@
+package demo.org.neo4j.graphql
+
+import org.junit.jupiter.api.TestFactory
+import org.neo4j.graphql.utils.CypherTestSuite
+
+class IssuesTests {
+
+    @TestFactory
+    fun `github #85`() = CypherTestSuite("issues/gh-85.adoc").generateTests()
+}
